// src/types/components/atoms.tsx
export interface ICTAtomComponentProps {
  children?: React.ReactNode;
  padding?: number | null | undefined;
  margin?: number | null | undefined;
  paddingBottom?: number | null | undefined;
  paddingTop?: number | null | undefined;
  paddingLeft?: number | null | undefined;
  paddingRight?: number | null | undefined;
  marginBottom?: number | null | undefined;
  marginTop?: number | null | undefined;
  marginLeft?: number | null | undefined;
  marginRight?: number | null | undefined;
  flex?: number | null | undefined;
  flexDirection?: string | null | undefined;
  alignItems?: string | null | undefined;
  justifyContent?: string | null | undefined;
  backgroundColor?: string | null | undefined;
  fontWeight?: string | number | null | undefined;
  fontSize?: string | number | null | undefined;
  color?: string | null | undefined;
  style?: any;
}